fix(blocks-react): reach the same stale-sheet verdict from both entry points
The migration check ran only in the exported read path, so a page rendered
through PageRenderer kept the rules a migration made stale — the primary entry
point, and the divergence the shared verdict helpers exist to prevent.

The check moves beside those helpers and both callers ask it.
