+++ packages/blocks-react/src/read-page.ts
@@ -18,7 +18,6 @@
  *
  * @module read-page
  */
-import { nodeAtPointer } from "@nextlyhq/blocks-engine";
 import type {
   BlockDocument,
   RemotePatternInput,
@@ -33,10 +32,10 @@ import { prepareDocumentReadStages, readingViewOf } from "./prepare-document";
 import type { BlockResolver } from "./resolver";
 import type { PageStyles } from "./styles";
 import {
-  drawlessTestFor,
   effectiveCompile,
   gatedMapCoversPrunedNodes,
   hasDuplicateNodeIds,
+  migrationChangedWhatDraws,
   readableGatedRules,
   resolvePageStyles,
 } from "./styles";
@@ -77,46 +76,6 @@ export interface PreparedPage {
   document: BlockDocument | null;
   /** The stylesheet for that tree. */
   styles: PageStyles;
-}
-
-/**
- * Whether migration turned a node that DREW into one that draws nothing.
- *
- * The stored sheet was compiled while the node still drew, so its rules sit in
- * the main `css` with no gated entry, and every later pass agrees the node is
- * present and registered. Nothing else in this function can see it: the
- * drawless predicate decides which per-node entries to APPEND and cannot
- * withdraw a rule already embedded in the sheet, and the stage comparisons all
- * read equal because no pass REMOVED anything. Only a recompile drops it.
- *
- * Asked ONLY of the nodes the engine reported rewriting. The predicate runs a
- * block's own declaration, so asking it of every node on every read would put
- * plugin code in the path of the ordinary case, where nothing migrated and the
- * answer cannot have changed. That list is empty for any page whose nodes are
- * already current, which is nearly all of them, and this returns before walking
- * anything.
- *
- * The reverse flip needs no handling: a node that starts drawless and begins to
- * draw has no rules in the stored sheet to be stale, and the recompile that
- * gives it some is triggered by the sheet not describing it.
- */
-function migrationChangedWhatDraws(
-  stages: DocumentReadStages,
-  resolver: BlockResolver
-): boolean {
-  if (stages.rewritten.length === 0) return false;
-  const drawsNothing = drawlessTestFor(resolver);
-  // The engine's reported pointers are RESOLVED rather than matched against
-  // paths rebuilt here. Rebuilding them would be a second encoder of the same
-  // format: it agrees on every ordinary slot name and diverges on the ones the
-  // escaping exists for, so the mismatch appears only for the documents this
-  // check most needs to read.
-  return stages.rewritten.some(entry => {
-    const before = nodeAtPointer(stages.sanitized, entry.path);
-    const after = nodeAtPointer(stages.migrated, entry.path);
-    if (before === undefined || after === undefined) return false;
-    return !drawsNothing(before) && drawsNothing(after);
-  });
 }
 
 /**
